Conflicts:
	Projects/MacOSFramework/Framework.xcodeproj/project.pbxproj
	Projects/Win32Framework/Win32Framework.vcxproj
	Projects/Win32Framework/Win32Framework.vcxproj.filters
	Sources/Internal/Render/Highlevel/RenderPass.cpp
	Tools/ResourceEditor/Classes/Qt/Main/mainwindow.cpp
	Tools/ResourceEditor/Data/Materials/Shaders/Default/materials.fsh
	Tools/ResourceEditor/Data/Materials/Shaders/Default/materials.vsh
	Tools/ResourceEditor/Data/Materials/assignable.yaml

Former-commit-id: afbada8af0ff2e2fee7bf87f23aec07e1befcb5c

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem/HeightmapProxy.cpp
@@ -39,7 +39,8 @@ void HeightmapProxy::UpdateRect(const DAVA::Rect &rect)
 {
 	int32 size = Size();
 
-	Rect bounds(0.f, 0.f, size, size);
+	Rect bounds(0.f, 0.f, (float32)size, (float32)size);
+
 	changedRect = rect;
 	bounds.ClampToRect(changedRect);
